Use pointer-sized word for limbs by default
All tests now should work for both 32-bit and 64-bit words. These should
be refined a bit further and be arbitrary, however.

There are some pending issues for smaller word sizes (e.g. 16-bit) on
the set/to functions but otherwise the main core seems to work as
expected according to the bench tests. Much slower though, so the
use-case of small limbs is mostly non-existent.

A compiler bigint bug to fix; `comptime a >> 64` case fails. We work around
this for the moment by splitting the shifts in two.

Closes #4.