Pre-decrement the divisors in facdiv
Since the value `c` was post-incremented while building the factorial
value, it needs to be pre-decremented when dividing it down.  All of
the implementations except Python were printing `0`, because they
started their division with the value 20001.

However, the Python implementation was also missing the final factor of
20000, because it only multiplied `while c < target`.  This has also
been corrected, and all implementations now also print their factorial
value to make sure they're the same.

(None of this makes much difference to the benchmark numbers though.)