Update for value-type parameters
This also adds `Int.initFixed` which allows initializing a read-only Int
type with a specific set of limbs values. This is used internally for
constant arguments to avoid allocations. Ideally this would be done
using FixedBufferAllocator and the like but the allocation strategies
performed allocate quite a lot more memory than we expect. We should
however revisit this approach as it should be the preferred.